feat: add buttons to clear and apply filters in FilterSidebar
The animation of the buttons sliding up is actually done by making them
static positioned and decreasing the height of the filters container
element.
the FliterProducts component now has a useEffect with a timeout to know
when the animation transition of the sidebar is done and update the state
of it accordingly and use this state to decrease the height of the above
element when the time is right.