feat: add color filters in ProductDetails with upper and sole SoleColors
These filters are dynamic that if you chose a sole color it will set the
sole color for all the upperColors accordingly if it's available for that
upper color, and when you click the upper color it will use the locally
stored sole color that changed earlier to set it globally.

PS. this is unstyled and using a mix of dummy and real data.